Separate out the libgmt loading tests
